Feat: 매물 가격 필터 구현

+++ src/main/java/com/sido/backend/realestate/service/RealEstateService.java
@@ -7,7 +7,7 @@ import com.sido.backend.realestate.entity.RealEstate;
 
 public interface RealEstateService {
 	PageResponseDTO<RealEstateResponseDTO, RealEstate> getRealEstateList(int page, int listSize, String location,
-		String tradeType, Integer minPrice, Integer maxPrice);
+		String tradeType, Long minPrice, Long maxPrice);
 
 	RealEstateDetailResponseDTO getRealEstateDetail(Long id);
 }
