+++ src/main/java/com/sido/backend/realestate/service/RealEstateServiceImpl.java
@@ -30,7 +30,7 @@ public class RealEstateServiceImpl implements RealEstateService {
 
 	@Override
 	public PageResponseDTO<RealEstateResponseDTO, RealEstate> getRealEstateList(int page, int listSize, String address,
-		String tradeType, Integer minPrice, Integer maxPrice) {
+		String tradeType, Long minPrice, Long maxPrice) {
 		Slice<RealEstate> lists = realEstatesRepository.findRealEstatesDynamically(
 			address, tradeType, minPrice, maxPrice, PageRequest.of(page - 1, listSize, Sort.by(Sort.Order.desc("id"))));
 		return new PageResponseDTO<>(lists, this::toDTO);
